+++ core/src/vpn_manager.h
@@ -97,7 +97,6 @@ struct Vpn {
     VpnHandler handler = {};
     DeclPtr<VpnNetworkManager, &vpn_network_manager_destroy> network_manager{vpn_network_manager_get()};
     AutoPod<VpnUpstreamConfig, vpn_upstream_config_destroy> upstream_config;
-    std::vector<sockaddr_storage> relay_addresses;
     /** The endpoint the client is connected or trying to connect to */
     std::optional<SelectedEndpoint> selected_endpoint;
     bool network_changed_before_recovery = false;