+++ net/src/ping.cpp
@@ -31,17 +31,18 @@
 
 namespace ag {
 
-static ag::Logger g_logger{"PING"}; // NOLINT(cert-err58-cpp)
+static ag::Logger g_logger{"PING"}; // NOLINT(cert-err58-cpp,cppcoreguidelines-avoid-non-const-global-variables)
 
 #define log_ping(ping_, lvl_, fmt_, ...) lvl_##log(g_logger, "[{}] " fmt_, (ping_)->id, ##__VA_ARGS__)
 #define log_conn(ping_, conn_, lvl_, fmt_, ...)                                                                        \
-    log_ping(ping_, lvl_, "{}{} ({}{}) via {}: " fmt_, (ping_)->use_quic ? "udp://" : "tcp://",                        \
-            (conn_)->endpoint->name, (conn_)->use_relay ? "through relay " : "",                                       \
-            (conn_)->use_relay ? sockaddr_to_str((sockaddr *) &(ping_)->relay_address)                                 \
-                               : sockaddr_to_str((sockaddr *) &(conn_)->endpoint->address),                            \
+    log_ping(ping_, lvl_, "Round {}: {}{} ({}{}) via {}: " fmt_, (ping_)->rounds_started,                              \
+            (conn_)->use_quic ? "udp://" : "tcp://", (conn_)->endpoint->name,                                          \
+            (conn_)->relay_address.ss_family ? "through relay " : "",                                                  \
+            (conn_)->relay_address.ss_family ? sockaddr_to_str((sockaddr *) &(conn_)->relay_address)                   \
+                                             : sockaddr_to_str((sockaddr *) &(conn_)->endpoint->address),              \
             (conn_)->bound_if_name, ##__VA_ARGS__)
 
-static std::atomic_int g_next_id;
+static std::atomic_int g_next_id; // NOLINT(cppcoreguidelines-avoid-non-const-global-variables)
 
 using PingClock = std::chrono::high_resolution_clock;
 using std::chrono::duration_cast;
@@ -96,7 +97,7 @@ enum PingConnState {
 
 struct PingConn {
     AutoVpnEndpoint endpoint;
-    bool use_relay = false;
+    sockaddr_storage relay_address{};
     std::vector<uint8_t> hello;
     AutoFd fd;
     DeclPtr<event, &event_free> event;
@@ -106,6 +107,9 @@ struct PingConn {
     std::string bound_if_name;
     int socket_error = 0;
     PingConnState state = PCS_SYN_SENT;
+    bool use_quic = false;
+    uint32_t rounds_done = 0;
+    std::vector<sockaddr_storage> relay_addresses; // These are in reverse order compared to the ones in `PingInfo`.
 };
 
 struct Ping {
@@ -114,16 +118,15 @@ struct Ping {
     VpnEventLoop *loop;
     PingHandler handler;
 
-    std::list<PingConn> pending;
-    std::list<PingConn> inprogress;
-    std::list<PingConn> errors;
-    std::list<PingConn> done;
+    std::list<PingConn> pending; // Waiting to start connection.
+    std::list<PingConn> inprogress; // Connection started.
+    std::list<PingConn> done; // Ready for next round.
+    std::list<PingConn> report; // Ready to report.
 
     DeclPtr<event, &event_free> timer;
 
-    uint32_t rounds_failed;
+    uint32_t rounds_target;
     uint32_t rounds_started;
-    uint32_t rounds_total;
     uint32_t round_timeout_ms;
 
     event_loop::AutoTaskId prepare_task_id;
@@ -131,10 +134,7 @@ struct Ping {
     event_loop::AutoTaskId hello_task_id;
     event_loop::AutoTaskId report_task_id;
 
-    sockaddr_storage relay_address;
-
     bool have_round_winner;
-    bool use_quic;
     bool anti_dpi;
 };
 
@@ -162,12 +162,12 @@ static void on_event(evutil_socket_t fd, short, void *arg) {
     if (conn->socket_error != 0) {
         log_conn(self, conn, dbg, "Failed in state {}: ({}) {}", magic_enum::enum_name(conn->state), conn->socket_error,
                 evutil_socket_error_to_string(conn->socket_error));
-        goto error;
+        goto end_round;
     }
 
     switch (conn->state) {
     case PCS_SYN_SENT: // TCP connected, send hello.
-        assert(!self->use_quic);
+        assert(!conn->use_quic);
         assert(!conn->hello.empty());
 
         log_conn(self, conn, trace, "Connected");
@@ -180,11 +180,11 @@ static void on_event(evutil_socket_t fd, short, void *arg) {
                 conn->socket_error = evutil_socket_geterror(conn->fd.get());
                 log_conn(self, conn, dbg, "Failed to send hello fragment: ({}) {}", conn->socket_error,
                         evutil_socket_error_to_string(conn->socket_error));
-                goto error;
+                goto end_round;
             } else if (ret != (int) hello.size()) {
-                conn->socket_error = 0;
+                conn->socket_error = -1;
                 log_conn(self, conn, dbg, "Failed to send hello fragment: no buffer space");
-                goto error;
+                goto end_round;
             }
 
             timeval tv = ms_to_timeval(DPI_COOLDOWN_TIME.count());
@@ -192,9 +192,9 @@ static void on_event(evutil_socket_t fd, short, void *arg) {
             if (0 != event_assign(conn->event.get(), vpn_event_loop_get_base(self->loop),
                                   conn->fd.get(), EV_TIMEOUT, on_event, self)
                     || 0 != event_add(conn->event.get(), &tv)) {
-                conn->socket_error = 0;
+                conn->socket_error = -1;
                 log_conn(self, conn, dbg, "Failed to assign/add event to wait for hello fragment delay");
-                goto error;
+                goto end_round;
             }
             // clang-format on
 
@@ -214,19 +214,19 @@ static void on_event(evutil_socket_t fd, short, void *arg) {
             conn->socket_error = evutil_socket_geterror(conn->fd.get());
             log_conn(self, conn, dbg, "Failed to send hello: ({}) {}", conn->socket_error,
                     evutil_socket_error_to_string(conn->socket_error));
-            goto error;
+            goto end_round;
         } else if (ret != (int) hello.size()) {
-            conn->socket_error = 0;
+            conn->socket_error = -1;
             log_conn(self, conn, dbg, "Failed to send hello: no buffer space");
-            goto error;
+            goto end_round;
         }
 
         // clang-format off
         if (0 != event_assign(conn->event.get(), vpn_event_loop_get_base(self->loop), conn->fd.get(), EV_READ, on_event, self)
                 || 0 != event_add(conn->event.get(), nullptr)) {
-            conn->socket_error = 0;
+            conn->socket_error = -1;
             log_conn(self, conn, dbg, "Failed to assign/add event to wait for response");
-            goto error;
+            goto end_round;
         }
         // clang-format on
 
@@ -247,7 +247,7 @@ static void on_event(evutil_socket_t fd, short, void *arg) {
                 conn->socket_error = error;
                 log_conn(self, conn, trace, "Failed to receive response: ({}) {}", conn->socket_error,
                         evutil_socket_error_to_string(conn->socket_error));
-                goto error;
+                goto end_round;
 #ifdef _WIN32
             }
 #endif
@@ -255,37 +255,32 @@ static void on_event(evutil_socket_t fd, short, void *arg) {
             // Treat this as an error for the purpose of switching to relay IPs.
             conn->socket_error = ag::utils::AG_ECONNRESET;
             log_conn(self, conn, trace, "Failed to receive response: unexpected EOF");
-            goto error;
+            goto end_round;
         }
 
         log_conn(self, conn, trace, "Got response");
 
         conn->best_result_ms = std::min(dt_ms, conn->best_result_ms.value_or(INT_MAX));
-        conn->fd.reset();
-        conn->event.reset();
-        self->done.splice(self->done.end(), self->inprogress, conn);
 
         if (!std::exchange(self->have_round_winner, true)) {
             uint32_t to_ms = std::min(2 * dt_ms + MIN_SHORT_TIMEOUT_MS, MAX_SHORT_TIMEOUT_MS);
             auto to_tv = ms_to_timeval(to_ms);
             evtimer_add(self->timer.get(), &to_tv);
-            log_ping(self, dbg, "Reducing round timeout to {} ms", to_ms);
+            log_ping(self, dbg, "Round {}: timeout reduced to {} ms", self->rounds_started, to_ms);
         }
         break;
     }
     }
 
-    goto end_round;
-
-error:
+end_round:
     conn->fd.reset();
     conn->event.reset();
-    self->errors.splice(self->errors.end(), self->inprogress, conn);
 
-end_round:
+    self->done.splice(self->done.end(), self->inprogress, conn);
+
     // All done or errors.
     if (self->inprogress.empty() && self->pending.empty()) {
-        log_ping(self, dbg, "Completed round {} of {}", self->rounds_started, self->rounds_total);
+        log_ping(self, dbg, "Round {}: complete", self->rounds_started);
         evtimer_del(self->timer.get());
         self->prepare_task_id = event_loop::submit(self->loop,
                 {
@@ -304,7 +299,7 @@ static void on_timer(evutil_socket_t, short, void *arg) {
 
     assert(!self->report_task_id.has_value());
 
-    log_ping(self, dbg, "Round {} of {} timed out", self->rounds_started, self->rounds_total);
+    log_ping(self, dbg, "Round {}: timed out", self->rounds_started);
 
     self->pending.splice(self->pending.end(), self->inprogress);
     for (PingConn &ep : self->pending) {
@@ -328,8 +323,8 @@ static void on_timer(evutil_socket_t, short, void *arg) {
 }
 
 // Return 0 if connection started successfully (including if it is inprogress), errno (or equivalent) otherwise.
-static int xconnect(Ping *ping, const PingConn &conn) {
-    auto *dest = conn.use_relay ? (sockaddr *) &ping->relay_address : (sockaddr *) &conn.endpoint->address;
+static int xconnect(const PingConn &conn) {
+    auto *dest = conn.relay_address.ss_family ? (sockaddr *) &conn.relay_address : (sockaddr *) &conn.endpoint->address;
     if (0 == connect(conn.fd.get(), dest, (int) sockaddr_get_size(dest))) {
         return 0;
     }
@@ -342,8 +337,8 @@ static int xconnect(Ping *ping, const PingConn &conn) {
 }
 
 // Return 0 if initial packet was sent successfully, errno (or equivalent) otherwise.
-static int send_quic_initial(Ping *ping, const PingConn &conn) {
-    auto *dest = conn.use_relay ? (sockaddr *) &ping->relay_address : (sockaddr *) &conn.endpoint->address;
+static int send_quic_initial(const PingConn &conn) {
+    auto *dest = conn.relay_address.ss_family ? (sockaddr *) &conn.relay_address : (sockaddr *) &conn.endpoint->address;
     int dest_size = (int) sockaddr_get_size(dest);
     if (0 != connect(conn.fd.get(), dest, dest_size)) {
         return evutil_socket_geterror(conn.fd.get());
@@ -365,26 +360,27 @@ static void do_connect(void *arg) {
 
     log_conn(self, conn, trace, "Connecting");
     conn->started_at = PingClock::now();
-    conn->socket_error = self->use_quic ? send_quic_initial(self, *conn) : xconnect(self, *conn);
+    conn->socket_error = conn->use_quic ? send_quic_initial(*conn) : xconnect(*conn);
     if (conn->socket_error != 0) {
-        log_conn(self, conn, dbg, "Failed to {}: {}: ({}) {}", self->use_quic ? "send initial" : "connect",
-                self->use_quic ? "send_quic_initial" : "xconnect", conn->socket_error,
+        log_conn(self, conn, dbg, "Failed to {}: {}: ({}) {}", conn->use_quic ? "send initial" : "connect",
+                conn->use_quic ? "send_quic_initial" : "xconnect", conn->socket_error,
                 evutil_socket_error_to_string(conn->socket_error));
         goto error;
     }
     if (0 != event_add(conn->event.get(), nullptr)) {
         log_conn(self, conn, dbg, "Failed to add event");
+        conn->socket_error = -1;
         goto error;
     }
 
-    conn->state = self->use_quic ? PCS_HELLO_SENT : PCS_SYN_SENT;
+    conn->state = conn->use_quic ? PCS_HELLO_SENT : PCS_SYN_SENT;
     self->inprogress.splice(self->inprogress.end(), self->pending, conn);
     goto next;
 
 error:
     conn->fd.reset();
     conn->event.reset();
-    self->errors.splice(self->errors.end(), self->pending, conn);
+    self->report.splice(self->report.end(), self->pending, conn);
 
 next:
     if (!self->pending.empty()) {
@@ -397,7 +393,18 @@ next:
                                     do_connect(arg);
                                 },
                 },
-                Millis{1} /*ms to force libevent ot poll/select between connect callss*/);
+                Millis{1} /*force libevent to poll/select between connect calls*/);
+    } else if (self->inprogress.empty()) {
+        // All failed (some may have started and already finished). Run `do_prepare` to decide what to do next.
+        evtimer_del(self->timer.get());
+        self->prepare_task_id = event_loop::submit(self->loop,
+                {
+                        .arg = self,
+                        .action =
+                                [](void *arg, TaskId) {
+                                    do_prepare(arg);
+                                },
+                });
     }
 }
 
@@ -410,37 +417,25 @@ static void do_report(void *arg) {
     assert(!self->connect_task_id.has_value());
     assert(!self->prepare_task_id.has_value());
 
-    PingResult result{
-            .ping = self,
-            .status = PING_OK,
-    };
+    PingResult result{.ping = self};
 
-    if (!self->done.empty()) {
-        auto it = self->done.begin();
-        result.through_relay = it->use_relay;
-        result.endpoint = it->endpoint.get();
-        if (it->best_result_ms.has_value()) {
-            result.ms = it->best_result_ms.value();
-        } else {
-            result.status = PING_TIMEDOUT;
+    if (!self->report.empty()) {
+        auto it = self->report.begin();
+        if (it->relay_address.ss_family) {
+            result.relay_address = (sockaddr *) &it->relay_address;
         }
-        self->handler.func(self->handler.arg, &result);
-        self->done.pop_front();
-        goto schedule_next;
-    }
-
-    if (!self->errors.empty()) {
-        auto it = self->errors.begin();
-        result.through_relay = it->use_relay;
         result.endpoint = it->endpoint.get();
         if (it->best_result_ms.has_value()) {
+            result.status = PING_OK;
             result.ms = it->best_result_ms.value();
         } else {
-            result.status = PING_SOCKET_ERROR;
             result.socket_error = it->socket_error;
+            result.status = (it->socket_error == 0 || it->socket_error == ag::utils::AG_ETIMEDOUT) ? PING_TIMEDOUT
+                                                                                                   : PING_SOCKET_ERROR;
+            result.ms = -1;
         }
         self->handler.func(self->handler.arg, &result);
-        self->errors.pop_front();
+        self->report.pop_front();
         goto schedule_next;
     }
 
@@ -468,11 +463,37 @@ static void do_prepare(void *arg) {
     assert(!self->connect_task_id.has_value());
     assert(!self->report_task_id.has_value());
     assert(self->inprogress.empty());
-    assert(!self->pending.empty() ? (self->errors.empty() && self->done.empty())
-                                  : (!self->errors.empty() || !self->done.empty()));
+    assert(!self->pending.empty() ? (self->done.empty() && self->report.empty())
+                                  : (!self->done.empty() || !self->report.empty()));
 
-    if (self->rounds_total == self->rounds_started) {
-        log_ping(self, dbg, "Pinging done, reporting results", self->rounds_started, self->rounds_total);
+    for (auto conn = self->done.begin(); conn != self->done.end();) {
+        // If an error occurred and we don't have any results yet, try falling back to TLS or the next relay address.
+        // (If an error occurred and we DO have a result, the error is most probably sporadic).
+        if (conn->socket_error && !conn->best_result_ms.has_value()) {
+            if (conn->use_quic) { // Fallback from QUIC to TLS
+                conn->use_quic = false;
+                conn->hello.clear();
+            } else if (!conn->relay_addresses.empty()) { // Fallback to the next relay address
+                conn->relay_address = conn->relay_addresses.back();
+                conn->relay_addresses.pop_back();
+            } else {
+                goto increment_rounds;
+            }
+            ++conn;
+            continue;
+        }
+        increment_rounds:
+        if (++conn->rounds_done == self->rounds_target) {
+            self->report.splice(self->report.end(), self->done, conn++);
+        } else {
+            ++conn;
+        }
+    }
+
+    self->pending.splice(self->pending.end(), self->done);
+
+    if (self->pending.empty()) {
+        log_ping(self, dbg, "Pinging done, reporting results");
         self->timer.reset();
         self->report_task_id = event_loop::submit(self->loop,
                 {
@@ -488,48 +509,32 @@ static void do_prepare(void *arg) {
     ++self->rounds_started;
     self->have_round_winner = false;
 
-    bool recalculate_initial = false;
-
-    if (!self->use_quic && self->done.empty() && !self->errors.empty()
-            && ++self->rounds_failed == self->rounds_total - 1) {
-        self->use_quic = true;
-        recalculate_initial = true;
-    }
-
-    log_ping(self, dbg, "Starting round {} of {}", self->rounds_started, self->rounds_total);
-
-    self->pending.splice(self->pending.end(), self->errors);
-    self->pending.splice(self->pending.end(), self->done);
-
     auto tv = ms_to_timeval(self->round_timeout_ms);
     evtimer_add(self->timer.get(), &tv);
 
     for (auto conn = self->pending.begin(); conn != self->pending.end();) {
-        if (std::exchange(conn->socket_error, 0) && self->relay_address.ss_family) {
-            conn->use_relay = true;
-        }
-        const auto *dest = (sockaddr *) (conn->use_relay ? &self->relay_address : &conn->endpoint->address);
+        conn->socket_error = 0;
+        const auto *dest =
+                (sockaddr *) (conn->relay_address.ss_family ? &conn->relay_address : &conn->endpoint->address);
         // NOLINTNEXTLINE(*-narrowing-conversions)
-        conn->fd = AutoFd(socket(dest->sa_family, self->use_quic ? SOCK_DGRAM : SOCK_STREAM, 0));
+        conn->fd = AutoFd(socket(dest->sa_family, conn->use_quic ? SOCK_DGRAM : SOCK_STREAM, 0));
         if (!conn->fd.valid()) {
-            log_conn(self, conn, dbg, "Failed to create socket");
+            conn->socket_error = evutil_socket_geterror(conn->fd.get());
+            log_conn(self, conn, dbg, "Failed to create socket: ({})", conn->socket_error,
+                    evutil_socket_error_to_string(conn->socket_error));
             goto error;
         }
         if (0 != evutil_make_socket_nonblocking(conn->fd.get())) {
-            log_conn(self, conn, dbg, "Failed to make socket non-blocking");
+            conn->socket_error = evutil_socket_geterror(conn->fd.get());
+            log_conn(self, conn, dbg, "Failed to make socket non-blocking: ({}) {}", conn->socket_error,
+                    evutil_socket_error_to_string(conn->socket_error));
             goto error;
         }
-        // clang-format off
-        if (int nodelay = 1; !self->use_quic && 0 != setsockopt(conn->fd.get(), IPPROTO_TCP, TCP_NODELAY,
-                                                                (char *) &nodelay, (int) sizeof(nodelay))) {
-            log_conn(self, conn, dbg, "Failed to set TCP_NODELAY");
-            goto error;
-        }
-        // clang-format on
 #ifndef _WIN32
         if (conn->bound_if != 0) {
 #ifdef __MACH__
-            const sockaddr *dest = (sockaddr *) (conn->use_relay ? &self->relay_address : &conn->endpoint->address);
+            const sockaddr *dest =
+                    (sockaddr *) (conn->relay_address.ss_family ? &conn->relay_address : &conn->endpoint->address);
             int option = (dest->sa_family == AF_INET) ? IP_BOUND_IF : IPV6_BOUND_IF;
             int level = (dest->sa_family == AF_INET) ? IPPROTO_IP : IPPROTO_IPV6;
             int error = setsockopt(conn->fd.get(), level, option, &conn->bound_if, sizeof(conn->bound_if));
@@ -539,41 +544,51 @@ static void do_prepare(void *arg) {
 #endif // #ifdef __MACH__
             if (error) {
                 log_conn(self, conn, dbg, "Failed to bind socket to interface: ({}) {}", errno, strerror(errno));
+                conn->socket_error = error;
                 goto error;
             }
         }
 #else  // #ifndef _WIN32
         if (!vpn_win_socket_protect(conn->fd.get(), dest)) {
             log_conn(self, conn, dbg, "Failed to protect socket");
+            conn->socket_error = -1;
             goto error;
         }
 #endif // #ifndef _WIN32
-        if (!self->use_quic) {
-            // Send RST as soon as socket is closed.
+        if (!conn->use_quic) {
+            // Send RST as soon as socket is closed. Ignore error, this is not essential.
             linger linger_0 = {.l_onoff = 1, .l_linger = 0};
             setsockopt(conn->fd.get(), SOL_SOCKET, SO_LINGER, (char *) &linger_0, (int) sizeof(linger_0));
+            // Send data after each send() call.
+            int nodelay = 1;
+            if (0 != setsockopt(conn->fd.get(), IPPROTO_TCP, TCP_NODELAY, (char *) &nodelay, (int) sizeof(nodelay))) {
+                conn->socket_error = evutil_socket_geterror(conn->fd.get());
+                log_conn(self, conn, dbg, "Failed to set TCP_NODELAY: ({}) {}", conn->socket_error,
+                        evutil_socket_error_to_string(conn->socket_error));
+                goto error;
+            }
         }
         conn->event.reset(event_new(vpn_event_loop_get_base(self->loop), conn->fd.get(),
-                self->use_quic ? EV_READ : EV_WRITE, on_event, self));
+                conn->use_quic ? EV_READ : EV_WRITE, on_event, self));
         if (conn->event == nullptr) {
             log_conn(self, conn, dbg, "Failed to create event");
+            conn->socket_error = -1;
             goto error;
         }
-        if (conn->hello.empty() || recalculate_initial) {
-            conn->hello = self->use_quic ? prepare_quic_initial(conn->endpoint->name)
+        if (conn->hello.empty()) {
+            conn->hello = conn->use_quic ? prepare_quic_initial(conn->endpoint->name)
                                          : prepare_client_hello(conn->endpoint->name);
         }
         ++conn;
         continue;
     error:
         conn->fd.reset();
-        auto next = std::next(conn);
-        self->errors.splice(self->errors.end(), self->pending, conn);
-        conn = next;
+        self->report.splice(self->report.end(), self->pending, conn++);
     }
 
     if (self->pending.empty()) {
         // All errors, start next round or report result.
+        evtimer_del(self->timer.get());
         self->prepare_task_id = event_loop::submit(self->loop,
                 {
                         .arg = self,
@@ -610,51 +625,44 @@ Ping *ping_start(const PingInfo *info, PingHandler handler) {
 
     self->loop = info->loop;
     self->handler = handler;
-    self->use_quic = info->use_quic;
     self->anti_dpi = info->anti_dpi;
+    self->rounds_target = info->nrounds ? info->nrounds : DEFAULT_PING_ROUNDS;
 
-    self->rounds_total = info->nrounds ? info->nrounds : DEFAULT_PING_ROUNDS;
     self->round_timeout_ms = info->timeout_ms ? info->timeout_ms : DEFAULT_PING_TIMEOUT_MS;
-
-    self->round_timeout_ms /= self->rounds_total;
     self->timer.reset(evtimer_new(vpn_event_loop_get_base(self->loop), on_timer, self.get()));
 
-    assert(self->rounds_total > 0);
-    assert(self->round_timeout_ms > 0);
-
-    if (info->relay_address) {
-        self->relay_address = sockaddr_to_storage(info->relay_address);
-    }
-
     constexpr uint32_t DEFAULT_IF_IDX = 0;
-    std::span<uint32_t> interfaces = info->interfaces_to_query;
+    std::span<const uint32_t> interfaces = info->interfaces_to_query;
     if (interfaces.empty()) {
         interfaces = {(uint32_t *) &DEFAULT_IF_IDX, size_t(1)};
     }
-    for (const VpnEndpoint &dst : info->endpoints) {
+    for (const VpnEndpoint &endpoint : info->endpoints) {
         for (uint32_t bound_if : interfaces) {
-            PingConn &endpoint = self->pending.emplace_back();
-            endpoint.endpoint = vpn_endpoint_clone(&dst);
-            endpoint.bound_if = bound_if;
-            if (ag::utils::trim(safe_to_string_view(dst.name)).empty()) {
-                log_ping(self, warn, "Endpoint {} has no name", sockaddr_to_str((sockaddr *) &dst.address));
+            PingConn &conn = self->pending.emplace_back();
+            conn.endpoint = vpn_endpoint_clone(&endpoint);
+            conn.bound_if = bound_if;
+            conn.use_quic = info->use_quic;
+            conn.relay_addresses.insert(
+                    conn.relay_addresses.begin(), info->relay_addresses.rbegin(), info->relay_addresses.rend());
+            if (ag::utils::trim(safe_to_string_view(endpoint.name)).empty()) {
+                log_ping(self, warn, "Endpoint {} has no name", sockaddr_to_str((sockaddr *) &endpoint.address));
                 return nullptr;
             }
 
             char buf[IF_NAMESIZE]{};
             if (bound_if != 0) {
                 if (if_indextoname(bound_if, buf)) {
-                    endpoint.bound_if_name = buf;
+                    conn.bound_if_name = buf;
                 } else {
 #ifndef _WIN32
                     log_ping(self, dbg, "if_indextoname: ({}) {}", errno, strerror(errno));
 #else
                     log_ping(self, dbg, "if_indextoname failed");
 #endif
-                    endpoint.bound_if_name = "(unknown)";
+                    conn.bound_if_name = "(unknown)";
                 }
             } else {
-                endpoint.bound_if_name = "(default)";
+                conn.bound_if_name = "(default)";
             }
         }
     }